prevent windows dialog functions from changing the PWD
"This flag is ineffective for GetOpenFileName." really amazing documentation from microsoft, we love to see it

Not setting this flag will change the PWD to whatever folder the user last selected, which also locks that directory and prevent any deletion or rename long after the file picker dialog was closed. No idea why this is the default behavior.